requirements.txt
asgiref==3.4.1
Django==3.2.8
psycopg2==2.9.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
djangorestframework
django-cors-headers
